refactor: organize components folder structure
